+++ CHANGELOG.md
@@ -16,6 +16,7 @@ You can also check the [release page](https://github.com/visualize-admin/visuali
   - Added support for a new dimension type - TemporalOrdinalDimension. This dimension can be used in all the places where you can use regular ordinal dimensions as well as to animate the chart
   - Scatterplot and Pie now animate updates
   - Axes now animate updates
+  - Tooltips for area, column and line charts now behave in more responsive way
   - Time filter now shows explicit `from` and `to` fields to filter by (and the brush snaps to values actually existing in the data)
   - Time filter now uses date picker for months and years 🗓️
   - Changed the aspect ratios of Pie and Scatterplot charts in the Editor mode, so they fit on a screen without the need to scroll